dnssd: windows: Greatly simplify code
The previous code was very complex as it was handling A and AAAA records
as well as SRV records, supposedly to support very old versions of Avahi
which sent incomplete or broken SRV records.

Since the version of Avahi we currently use seems to send us SRV records
containing all the information we need, we can't reproduce the original
issue - and therefore the code can be simplified greatly.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>